Refactor module to remove deprecated code and simplify structure
