+++ neon_diana_utils/kubernetes_utils.py
@@ -24,17 +24,47 @@
 # NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 # SOFTWARE,  EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-import os
-import sys
-import unittest
+import yaml
 
-sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
-from neon_diana_utils.rabbitmq_api import RabbitMQAPI
+from typing import Optional
+from os import getenv
+from os.path import join, expanduser
 
-
-class TestRabbitMQAPI(unittest.TestCase):
-    pass
+from neon_diana_utils.configuration import validate_output_path
 
 
-if __name__ == '__main__':
-    unittest.main()
+def create_github_secret(username: str, token: str,
+                         output_file: Optional[str] = None) -> str:
+    """
+    Generate a Kubernetes Secret to authenticate to GitHub for image pulls
+    :param username: GitHub username
+    :param token: GitHub token with read_packages permission
+    :param output_file: output file to write
+    :returns: path to written Kubernetes config file
+    """
+    import json
+    from base64 import b64encode
+    output_file = output_file or join(getenv("NEON_CONFIG_PATH",
+                                             "~/.config/neon"),
+                                      f"k8s_secret_github.yml")
+    output_file = expanduser(output_file)
+    if not validate_output_path(output_file):
+        raise FileExistsError(output_file)
+    encoded_auth = b64encode(f"{username}:{token}".encode())
+    auth_dict = {"auths": {"ghcr.io": {"auth": encoded_auth.decode()}}}
+    auth_str = json.dumps(auth_dict)
+    encoded_config = b64encode(auth_str.encode())
+    secret_spec = {
+        "kind": "Secret",
+        "type": "kubernetes.io/dockerconfigjson",
+        "apiVersion": "v1",
+        "metadata": {
+            "name": "github-auth"
+        },
+        "data": {
+            ".dockerconfigjson": encoded_config.decode()
+        }
+    }
+    with open(output_file, 'w+') as f:
+        yaml.dump(secret_spec, f)
+    return output_file
