--- neon_diana_utils/templates/docker-compose.yml
@@ -1,6 +1,0 @@
-version: "3.1"
-
-networks:
-  mq-backend:
-
-volumes: {}
