--- neon_diana_utils/templates/k8s_config_map.yml
@@ -1,5 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: ""
-data: {}
