--- neon_diana_utils/templates/k8s_secret.yml
@@ -1,6 +1,0 @@
-apiVersion: v1
-kind: Secret
-metadata:
-  name: ""
-type: Opaque
-stringData: {}
