--- neon_diana_utils/templates/kubernetes.yml
@@ -1,10 +1,0 @@
-apiVersion: v1
-kind: List
-metadata: {}
-items:
-  - kind: Namespace
-    apiVersion: v1
-    metadata:
-      name: ${MQ_NAMESPACE}
-      labels:
-        name: ${MQ_NAMESPACE}
