--- tests/config/ngi_local_conf.yml
@@ -1,239 +1,0 @@
-prefFlags:
-  codeSource: git
-  devMode: false
-  autoStart: false
-  autoUpdate: false
-  notifyRelease: true
-  showDemo: true
-  optDiags: false
-  metrics: false
-  saveAudio: false
-  saveText: true
-  # TODO: These should go somewhere in HAL/Enclosure
-  localGui: true
-  guiEvents: true
-
-interface:
-# True, False
-  display_neon_brain: false
-  wake_word_enabled: true
-  clap_commands_enabled: false
-  blink_commands_enabled: false
-  random_dialog_enabled: false
-  confirm_listening: true
-  mute_on_listen: false
-  # True=Muted, False=lowered vol
-  use_hesitation: false
-
-devVars:
-  devName:
-  devType: generic
-  # generic, server, pi, neonK, neonX, neonPi, neonAlpha, neonU, mycroft_mark_2, ubuntu, linux
-  version:
-  touchDev:
-  micDev:
-  camDev: 0
-  soundDev:
-  defaultVolume: 60
-  defaultMicVolume: 100
-  installUser:
-  # TODO: Depreciate installUser
-
-gestures:
-  clapThreshold: 10.0e10
-
-audioService:
-  backends:
-    local: {type: simple, active: true}
-    vlc: {type: vlc, active: true, duck: true}
-  defaultBackend: vlc
-  debug: true
-
-padatious:
-  intent_cache: ~/.neon/intent_cache
-  train_delay: 4
-  single_thread: false
-
-websocket:
-  host: 0.0.0.0
-  port: 8181
-  route: /core
-  ssl: false
-  allow_self_signed: true
-  ssl_cert:
-  ssl_key:
-
-gui:
-  # TODO: This should be built so lang, enclosure are read from other places
-  # Host information for the gui server
-  lang: en-us
-  enclosure: generic
-  host: 0.0.0.0
-  port: 18181
-  route: /gui
-  ssl: false
-  resource_root: https://0000.us/klatchat/app/files/neon_qml
-
-hotwords:
-  Hey Neon: {module: ovos-ww-plugin-pocketsphinx, phonemes: HH EY . N IY AA N ., threshold: 1e-20,
-    lang: en-us, sample_rate: 16000, listen: true, sound: snd/start_listening.wav,
-    local_model_file: None}
-
-listener:
-  sample_rate: 16000
-  channels: 1
-  wake_word_upload: {disable: true, url: https://training.mycroft.ai/precise/upload}
-  mute_during_output: true
-  duck_while_listening: 0.3
-  dev_index:
-  phoneme_duration: 120
-  multiplier: 1.0
-  energy_ratio: 1.5
-  wake_word: hey neon
-  stand_up_word: wake up
-  recording_timeout: 10.0,
-  recording_timeout_with_silence: 3.0
-
-skills:
-  auto_update: true
-  install_essential: true
-  install_default: true
-  wait_for_internet: true
-  debug: true
-  blacklist: []
-  priority: [about.neon.neongeckocom, date-time.neon.neongeckocom, skill-weather.neongeckocom]
-  essential_skills: []
-  default_skills: https://raw.githubusercontent.com/NeonGeckoCom/neon-skills-submodules/master/DEFAULT-SKILLS
-  skill_manager: osm
-  # recommended osm, optional msm
-  appstore_sync_interval: 6.0
-  # time between server syncs in hours
-  auto_update_interval: 24.0
-  # time between automatic skill updates in hours
-  msm_ver: false
-  repo_url: https://github.com/MycroftAI/mycroft-skills
-  repo_branch: '18.08'
-  data_dir: ~/.neon/msm
-  neon_token:
-
-audio_parsers:
-  blacklist: [gender]
-
-session:
-  ttl: 180
-tts:
-  module: amazon
-  amazon: {region: us-west-2, aws_access_key_id: '', aws_secret_access_key: ''}
-  mozilla: {request_url: http://0.0.0.0:5002/api/tts?}
-  mimic: {voice: ap}
-  mimic2: {lang: en-us, url: https://mimic-api.mycroft.ai/synthesize?text=}
-stt:
-  module: google_cloud_streaming
-
-language:
-  boost: false
-  core_lang: en-us
-  translation_module: libretranslate_plug
-  detection_module: libretranslate_detection_plug
-  libretranslate: {libretranslate_host: https://translate.neon.ai:5000}
-
-logs:
-  blacklist: [enclosure.mouth.viseme, enclosure.mouth.display]
-
-device:
-  mac: 00:00:00:00:00:00
-  ip4: 127.0.0.1
-  ip6: ::1
-  ver: ''
-
-api:
-  url: https://api.mycroft.ai
-  version: v1
-  update: true
-  disabled: true
-  sync_skill_settings: false
-
-# Below are (mostly) depreciated DM
-
-remoteVars:
-  coreGit:
-  coreBranch:
-  skillsGit:
-  skillsBranch:
-  guiGit:
-  guiBranch:
-  url: https://api.mycroft.ai
-  ver: v1
-  remoteHost: 167.172.112.7
-  enableConnection: true
-
-gnome:
-  favApps:
-  appFolders: "['accessories', 'office', 'Neongecko']"
-
-dirVars:
-  coreDir: /opt/neon
-  # coreDir is depreciated for packaged Neon Installations
-  rootDir: ~/.local/share/neon
-  confDir: ~/.config/neon
-  cacheDir: ~/.cache/neon
-  skillsDir: ~/.local/share/neon/skills
-  ngiDir: /opt/neon/NGI
-  # ngiDir is depreciated for packaged Neon Installations
-  guiDir:
-  tempDir: /tmp/neon
-  docsDir: ~/Documents/NeonGecko
-  diagsDir: ~/Documents/NeonGecko/Diagnostics
-  ipcDir: /tmp/neon/ipc
-  # Changes to ipcDir must also be made in NGI/utilities/configHelper
-  musicDir: ~/Music
-  videoDir: ~/Videos/NeonGecko
-  picsDir: ~/Pictures/NeonGecko
-  logsDir: /var/log/mycroft
-  repoDir: ~/.neon/skills-repo
-  padatiousDir: ~/.neon/intent_cache
-
-# File paths should be absolute (or ~/ relative) or relative to /res/ (see resolve_resource_file)
-fileVars:
-  sshKey: ~/.ssh/id_rsa
-  notify: snd/loaded.wav
-
-sounds:
-  startListening: snd/start_listening.wav
-  endListening: snd/end_listening.wav
-  acknowledge: snd/acknowledge.mp3
-
-ttsOpts: {None: '', 'Chinese, Mandarin': zh-ZH, Danish: da-DK, Dutch: nl-NL, 'English, Australian': en-AU,
-  'English, British': en-GB, 'English, Indian': en-IN, 'English, US': en-US, 'English, Welsh': en-GB-WLS,
-  French: fr-FR, 'French, Canadian': fr-CA, Hindi: hi-IN, German: de-DE, Icelandic: is-IS,
-  Italian: it-IT, Japanese: ja-JP, Korean: ko-KR, Norwegian: nb-NO, Polish: pl-PL,
-  'Portuguese, Brazilian': pt-BR, 'Portuguese, European': pt-PT, Romanian: ro-RO,
-  Russian: ru-RU, 'Spanish, European': es-ES, 'Spanish, Mexican': es-MX, 'Spanish, US': es-US,
-  Swedish: sv-SE, Turkish: tr-TR, Welsh: cy-GB}
-
-sttOpts: {Deutsch (Deutschland): de-DE, English (United States): en-US, Español (España): es-ES,
-  Español (México): es-MX, Français (Canada): fr-CA, Français (France): fr-FR, Italiano (Italia): it-IT,
-  Português (Portugal): pt-PT}
-
-sttSpokenOpts: {Afrikaans: af-ZA, Amharic: am-ET, Armenian: hy-AM, Azerbaijani: az-AZ,
-  Indonesian: id-ID, Malay: ms-MY, Bengali: bn-BD, Catalan: ca-ES, Czech: cs-CZ, Danish: da-DK,
-  German: de-DE, 'English, British': en-GB, 'English, American': en-US, Spanish: es-ES,
-  'Spanish, American': es-US, 'Spanish, Mexican': es-MX, Basque: eu-ES, Filipino: fil-PH,
-  'French, Canadian': fr-CA, French: fr-FR, Galician: gl-ES, Georgian: ka-GE, Gujarati: gu-IN,
-  Croatian: hr-HR, Zulu: zu-ZA, Icelandic: is-IS, Italian: it-IT, Javanese: jv-ID,
-  Kannada: kn-IN, Khmer: km-KH, Lao: lo-LA, Latvian: lv-LV, Lithuanian: lt-LT, Hungarian: hu-HU,
-  Malayalam: ml-IN, Marathi: mr-IN, Dutch: nl-NL, Nepali: ne-NP, Norwegian Bokmål: nb-NO,
-  Polish: pl-PL, 'Portuguese, Brazilian': pt-BR, Portuguese: pt-PT, Romanian: ro-RO,
-  Sinhala: si-LK, Slovak: sk-SK, Slovenian: sl-SI, Sundanese: su-ID, Swahili: sw-TZ,
-  Finnish: fi-FI, Swedish: sv-SE, Tamil: ta-IN, Telugu: te-IN, Vietnamese: vi-VN,
-  Turkish: tr-TR, Urdu: ur-PK, Greek: el-GR, Bulgarian: bg-BG, Russian: ru-RU, Serbian: sr-RS,
-  Ukrainian: uk-UA, Hebrew: he-IL, Arabic: ar-IL, Persian: fa-IR, Hindi: hi-IN, Thai: th-TH,
-  Korean: ko-KR, Taiwanese: zh-TW, 'Chinese, Cantonese': yue-Hant-HK, Japanese: ja-JP,
-  'Chinese, Mandarin': zh}
-
-MQ:
-  server: 0.0.0.0
-  users:
-    mq_handler:
-      user: neon_api_utils
-      password: Klatchat2021
