--- tests/config/valid_docker-compose-services.json
@@ -1,41 +1,0 @@
-{
-  "neon-metrics-service": {
-    "container_name": "neon-metrics-service",
-    "hostname": "neon-metrics-service",
-    "image": "ghcr.io/neongeckocom/neon-metrics-service:dev",
-    "restart": "unless-stopped",
-    "links": [
-      "neon-rabbitmq"
-    ],
-    "volumes": [
-      "host_path:/config",
-      "host_path_metrics:/metrics"
-    ]
-  },
-  "neon-rabbitmq": {
-    "container_name": "neon-rabbitmq",
-    "hostname": "neon-rabbitmq",
-    "image": "rabbitmq:3-management",
-    "restart": "unless-stopped",
-    "ports": [
-      "15672:15672",
-      "5672:5672"
-    ],
-    "volumes": [
-      "host_path/rabbitmq.conf:/etc/rabbitmq/rabbitmq.conf",
-      "host_path:/config"
-    ]
-  },
-  "neon-api-proxy": {
-    "container_name": "neon-api-proxy",
-    "hostname": "neon-api-proxy",
-    "image": "ghcr.io/neongeckocom/neon-api-proxy:dev",
-    "restart": "unless-stopped",
-    "links": [
-      "neon-rabbitmq"
-    ],
-    "volumes": [
-      "host_path:/config"
-    ]
-  }
-}
