--- tests/config/valid_rabbitmq.json
@@ -1,105 +1,0 @@
-{
-  "rabbit_version": "3.9.9",
-  "rabbitmq_version": "3.9.9",
-  "product_name": "RabbitMQ",
-  "product_version": "3.9.9",
-  "users": [
-    {
-      "name": "admin",
-      "password_hash": "wf7UYT6N8MZOD4LAQIs9CrUIG4qZ9bK0/OviUS2y7uHBrhPZ",
-      "hashing_algorithm": "rabbit_password_hashing_sha256",
-      "tags": [
-        "administrator"
-      ],
-      "limits": {}
-    },
-    {
-      "name": "neon_api_utils",
-      "password_hash": "S8JsaNbG7436awNWb1XyehfVcueqOvJU5ecEimzbcVCQt+d3",
-      "hashing_algorithm": "rabbit_password_hashing_sha256",
-      "tags": [],
-      "limits": {}
-    },
-    {
-      "name": "neon_metrics",
-      "password_hash": "7SgPQoQRcOSMrRCMZotCUu9zeKBK+z3YtBcG7jZ+bS69GoFa",
-      "hashing_algorithm": "rabbit_password_hashing_sha256",
-      "tags": [],
-      "limits": {}
-    },
-    {
-      "name": "neon_api",
-      "password_hash": "Ua92n5KKsUtIY49mlEnZLSnUtM/9Is6McCCCcP2qMFXUk0os",
-      "hashing_algorithm": "rabbit_password_hashing_sha256",
-      "tags": [],
-      "limits": {}
-    }
-  ],
-  "vhosts": [
-    {
-      "name": "/"
-    },
-    {
-      "name": "/neon_metrics"
-    },
-    {
-      "name": "/neon_api"
-    }
-  ],
-  "permissions": [
-    {
-      "user": "admin",
-      "vhost": "/neon_metrics",
-      "configure": ".*",
-      "write": ".*",
-      "read": ".*"
-    },
-    {
-      "user": "admin",
-      "vhost": "/neon_api",
-      "configure": ".*",
-      "write": ".*",
-      "read": ".*"
-    },
-    {
-      "user": "neon_api_utils",
-      "vhost": "/neon_metrics",
-      "configure": ".*",
-      "write": ".*",
-      "read": ""
-    },
-    {
-      "user": "neon_metrics",
-      "vhost": "/neon_metrics",
-      "configure": ".*",
-      "write": ".*",
-      "read": ".*"
-    },
-    {
-      "user": "neon_api_utils",
-      "vhost": "/neon_api",
-      "configure": "./*",
-      "write": "./*",
-      "read": "./*"
-    },
-    {
-      "user": "neon_api",
-      "vhost": "/neon_api",
-      "configure": ".*",
-      "write": ".*",
-      "read": ".*"
-    }
-  ],
-  "topic_permissions": [],
-  "parameters": [],
-  "global_parameters": [
-    {
-      "name": "internal_cluster_id",
-      "value": "rabbitmq-cluster-id-WZ_YB7W-2OfQZV_NAU1_kg"
-    }
-  ],
-  "policies": [],
-  "queues": [],
-  "exchanges": [],
-  "bindings": []
-}
