--- tests/test_diana_services_config/.env
@@ -1,2 +1,0 @@
-NEON_CONFIG_PATH=/home/d_mcknight/PycharmProjects/_.API/neon-diana-backend/tests/test_diana_services_config
-NEON_METRIC_PATH=/home/d_mcknight/PycharmProjects/_.API/neon-diana-backend/tests/test_diana_services_config/metrics
