--- tests/test_diana_services_config/docker-compose.yml
@@ -1,37 +1,0 @@
-version: '3.1'
-networks:
-  mq_backend:
-services:
-  neon-rabbitmq:
-    container_name: neon-rabbitmq
-    image: rabbitmq:3-management
-    restart: always
-    networks:
-    - mq_backend
-    ports:
-    - 15672:15672
-    - 5672:5672
-    volumes:
-      - ${NEON_CONFIG_PATH}/rabbitmq.conf:/etc/rabbitmq/rabbitmq.conf
-      - ${NEON_CONFIG_PATH}:/config
-  neon-metrics-service:
-    container_name: neon-metrics-service
-    image: ghcr.io/neongeckocom/neon_metrics_service:dev
-    restart: on-failure
-    networks:
-    - mq_backend
-    volumes:
-    - ${NEON_CONFIG_PATH}:/config
-    - ${NEON_CONFIG_PATH}/metrics:/metrics
-    labels:
-      kompose.service.type: headless
-  neon-api-proxy:
-    container_name: neon-api-proxy
-    image: ghcr.io/neongeckocom/neon_api_proxy:dev
-    restart: on-failure
-    networks:
-    - mq_backend
-    volumes:
-    - ${NEON_CONFIG_PATH}:/config
-    labels:
-      kompose.service.type: headless
