--- tests/test_diana_services_config/mq_config.json
@@ -1,13 +1,0 @@
-{
-  "server": "neon-rabbitmq",
-  "users": {
-    "neon_metrics_connector": {
-      "user": "neon_metrics",
-      "password": "0kUv28tVK0Omw3H4HOq7AqnDERoUqPmWx3-vdsLjmRQ"
-    },
-    "neon_api_connector": {
-      "user": "neon_api",
-      "password": "5akSUyWvSEV62ymdc6DvADUEnx46W5E8vZFmFjYbYGI"
-    }
-  }
-}
