--- tests/test_diana_services_config/ngi_auth_vars.yml
@@ -1,3 +1,0 @@
-track_my_brands: {}
-emails: {}
-api_services: {}
