--- tests/test_diana_services_config/rabbitmq.conf
@@ -1,1 +1,0 @@
-load_definitions = /config/rabbit_mq_config.json
